requirements.txt
mysql-connector-python>=8.0.0
sentence-transformers>=2.2.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.20.0
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
